Use regex for more robust query handling
