+++ packages/vite/src/node/plugins/asset.ts
@@ -30,11 +30,15 @@ import {
   withTrailingSlash,
 } from '../../shared/utils'
 import type { Environment } from '../environment'
+import { svgToDataURL } from './svg'
 
 // referenceId is base64url but replaces - with $
 export const assetUrlRE = /__VITE_ASSET__([\w$]+)__(?:\$_(.*?)__)?/g
 
 const jsSourceMapRE = /\.[cm]?js\.map$/
+
+const noInlineRE = /[?&]no-inline\b/
+const inlineRE = /[?&]inline\b/
 
 const assetCache = new WeakMap<Environment, Map<string, string>>()
 
@@ -414,7 +418,8 @@ const shouldInline = (
 ): boolean => {
   const environment = pluginContext.environment
   const { assetsInlineLimit } = environment.config.build
-  if (id.endsWith('?no-inline')) return false
+  if (noInlineRE.test(id)) return false
+  if (inlineRE.test(id)) return true
   if (environment.config.build.lib) return true
   if (pluginContext.getModuleInfo(id)?.isEntry) return false
   if (forceInline !== undefined) return forceInline
@@ -430,36 +435,4 @@ const shouldInline = (
   // Don't inline SVG with fragments, as they are meant to be reused
   if (file.endsWith('.svg') && id.includes('#')) return false
   return content.length < limit && !isGitLfsPlaceholder(content)
-}
-
-const nestedQuotesRE = /"[^"']*'[^"]*"|'[^'"]*"[^']*'/
-
-// Inspired by https://github.com/iconify/iconify/blob/main/packages/utils/src/svg/url.ts
-function svgToDataURL(content: Buffer): string {
-  const stringContent = content.toString()
-  // If the SVG contains some text or HTML, any transformation is unsafe, and given that double quotes would then
-  // need to be escaped, the gain to use a data URI would be ridiculous if not negative
-  if (
-    stringContent.includes('<text') ||
-    stringContent.includes('<foreignObject') ||
-    nestedQuotesRE.test(stringContent)
-  ) {
-    return `data:image/svg+xml;base64,${content.toString('base64')}`
-  } else {
-    return (
-      'data:image/svg+xml,' +
-      stringContent
-        .trim()
-        .replaceAll(/>\s+</g, '><')
-        .replaceAll('"', "'")
-        .replaceAll('%', '%25')
-        .replaceAll('#', '%23')
-        .replaceAll('<', '%3c')
-        .replaceAll('>', '%3e')
-        // Spaces are not valid in srcset it has some use cases
-        // it can make the uncompressed URI slightly higher than base64, but will compress way better
-        // https://github.com/vitejs/vite/pull/14643#issuecomment-1766288673
-        .replaceAll(/\s+/g, '%20')
-    )
-  }
 }
